entity 수정
