+++ issuestation/src/main/java/com/issuestation/Entity/Assignee.java
@@ -1,0 +1,22 @@
+package com.issuestation.Entity;
+import jakarta.persistence.*;
+import lombok.*;
+
+@Builder
+@Entity
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+public class Assignee {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "userid")
+    private UserEntity user;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "issueid")
+    private IssueEntity issue;
+}
