+++ issuestation/src/main/java/com/issuestation/Entity/Comment.java
@@ -1,0 +1,33 @@
+package com.issuestation.Entity;
+
+import com.issuestation.Entity.common.BaseTimeEntity;
+import com.issuestation.Entity.enums.commentTag;
+import jakarta.persistence.*;
+import jakarta.validation.constraints.NotNull;
+import lombok.*;
+
+@Builder
+@Entity
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+public class Comment extends BaseTimeEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @NotNull
+    private commentTag tag;
+
+    @NotNull
+    @Column(columnDefinition = "TEXT")
+    private String comment;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "userid")
+    private UserEntity user;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "issueid")
+    private IssueEntity issue;
+}
