+++ issuestation/src/main/java/com/issuestation/Entity/ReporterEntity.java
@@ -1,5 +1,6 @@
 package com.issuestation.Entity;
 
+import com.issuestation.Entity.Temp.ReporterId;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
