+++ issuestation/src/main/java/com/issuestation/Entity/Temp/AssigneeId.java
@@ -1,4 +1,4 @@
-package com.issuestation.Entity;
+package com.issuestation.Entity.Temp;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -6,10 +6,12 @@ import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
 import java.util.Objects;
+
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class ReporterId implements Serializable {
+public class AssigneeId implements Serializable {
     private int userid;
     private int issueid;
     private int pid;
@@ -19,8 +21,8 @@ public class ReporterId implements Serializable {
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
-        ReporterId reporterId = (ReporterId) o;
-        return (userid == reporterId.userid) && (issueid == reporterId.issueid);
+        AssigneeId assigneeId = (AssigneeId) o;
+        return (userid == assigneeId.userid) && (issueid == assigneeId.issueid);
     }
 
     @Override
