+++ issuestation/src/main/java/com/issuestation/Entity/Temp/IssueId.java
@@ -1,4 +1,4 @@
-package com.issuestation.Entity;
+package com.issuestation.Entity.Temp;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
