+++ issuestation/src/main/java/com/issuestation/Entity/common/BaseTimeEntity.java
@@ -22,5 +22,4 @@ public abstract class BaseTimeEntity {
     @LastModifiedDate
     @DateTimeFormat(pattern = "yyyy-MM-dd")
     private LocalDateTime moddate;
-
 }
