+++ issuestation/src/main/java/com/issuestation/Entity/enums/commentTag.java
@@ -1,0 +1,5 @@
+package com.issuestation.Entity.enums;
+
+public enum commentTag {
+    NEW, SETASSIGNEE, CLOSED, NORMAL
+}
