+++ issuestation/src/main/java/com/issuestation/Entity/enums/status.java
@@ -1,0 +1,5 @@
+package com.issuestation.Entity.enums;
+
+public enum status {
+    NEW, FIXED, CLOSED, REOPENED
+}
